// src/app/models/type-document.enum.ts
// models/type-document.enum.ts

export enum TypeDocument {
  PIECE_IDENTITE = 'PIECE_IDENTITE',
  DIPLOME_BAC = 'DIPLOME_BAC',
  DIPLOME_BAC_2 = 'DIPLOME_BAC_2',
  DIPLOME_BAC_3 = 'DIPLOME_BAC_3',
  CV = 'CV',
  LETTRE_MOTIVATION = 'LETTRE_MOTIVATION',
  JUSTIF_SITUATION = 'JUSTIF_SITUATION',
  JUSTIFICATIF = 'JUSTIFICATIF',
  PORTFOLIO = 'PORTFOLIO',
  ATTEST_RESP_CIVILE = 'ATTEST_RESP_CIVILE',
  AUTRE = 'AUTRE'
}